+++ lib/src/database/twonly.db.steps.dart
@@ -4303,6 +4303,391 @@ i1.GeneratedColumn<int> _column_206(String aliasedName) =>
     i1.GeneratedColumn<int>(
         'new_delete_messages_after_milliseconds', aliasedName, true,
         type: i1.DriftSqlType.int, $customConstraints: 'NULL');
+
+final class Schema9 extends i0.VersionedSchema {
+  Schema9({required super.database}) : super(version: 9);
+  @override
+  late final List<i1.DatabaseSchemaEntity> entities = [
+    contacts,
+    groups,
+    mediaFiles,
+    messages,
+    messageHistories,
+    reactions,
+    groupMembers,
+    receipts,
+    receivedReceipts,
+    signalIdentityKeyStores,
+    signalPreKeyStores,
+    signalSenderKeyStores,
+    signalSessionStores,
+    messageActions,
+    groupHistories,
+  ];
+  late final Shape22 contacts = Shape22(
+      source: i0.VersionedTable(
+        entityName: 'contacts',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(user_id)',
+        ],
+        columns: [
+          _column_106,
+          _column_107,
+          _column_108,
+          _column_109,
+          _column_110,
+          _column_111,
+          _column_112,
+          _column_113,
+          _column_114,
+          _column_115,
+          _column_116,
+          _column_117,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape23 groups = Shape23(
+      source: i0.VersionedTable(
+        entityName: 'groups',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(group_id)',
+        ],
+        columns: [
+          _column_119,
+          _column_120,
+          _column_121,
+          _column_122,
+          _column_123,
+          _column_124,
+          _column_125,
+          _column_126,
+          _column_127,
+          _column_128,
+          _column_129,
+          _column_130,
+          _column_131,
+          _column_132,
+          _column_133,
+          _column_134,
+          _column_118,
+          _column_135,
+          _column_136,
+          _column_137,
+          _column_138,
+          _column_139,
+          _column_140,
+          _column_141,
+          _column_142,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape36 mediaFiles = Shape36(
+      source: i0.VersionedTable(
+        entityName: 'media_files',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(media_id)',
+        ],
+        columns: [
+          _column_143,
+          _column_144,
+          _column_145,
+          _column_146,
+          _column_147,
+          _column_148,
+          _column_149,
+          _column_207,
+          _column_150,
+          _column_151,
+          _column_152,
+          _column_153,
+          _column_154,
+          _column_155,
+          _column_156,
+          _column_157,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape25 messages = Shape25(
+      source: i0.VersionedTable(
+        entityName: 'messages',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(message_id)',
+        ],
+        columns: [
+          _column_158,
+          _column_159,
+          _column_160,
+          _column_144,
+          _column_161,
+          _column_162,
+          _column_163,
+          _column_164,
+          _column_165,
+          _column_153,
+          _column_166,
+          _column_167,
+          _column_168,
+          _column_169,
+          _column_118,
+          _column_170,
+          _column_171,
+          _column_172,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape26 messageHistories = Shape26(
+      source: i0.VersionedTable(
+        entityName: 'message_histories',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [],
+        columns: [
+          _column_173,
+          _column_174,
+          _column_175,
+          _column_161,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape27 reactions = Shape27(
+      source: i0.VersionedTable(
+        entityName: 'reactions',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(message_id, sender_id, emoji)',
+        ],
+        columns: [
+          _column_174,
+          _column_176,
+          _column_177,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape28 groupMembers = Shape28(
+      source: i0.VersionedTable(
+        entityName: 'group_members',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(group_id, contact_id)',
+        ],
+        columns: [
+          _column_158,
+          _column_178,
+          _column_179,
+          _column_180,
+          _column_181,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape29 receipts = Shape29(
+      source: i0.VersionedTable(
+        entityName: 'receipts',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(receipt_id)',
+        ],
+        columns: [
+          _column_182,
+          _column_183,
+          _column_184,
+          _column_185,
+          _column_186,
+          _column_187,
+          _column_188,
+          _column_189,
+          _column_190,
+          _column_191,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape30 receivedReceipts = Shape30(
+      source: i0.VersionedTable(
+        entityName: 'received_receipts',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(receipt_id)',
+        ],
+        columns: [
+          _column_182,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape31 signalIdentityKeyStores = Shape31(
+      source: i0.VersionedTable(
+        entityName: 'signal_identity_key_stores',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(device_id, name)',
+        ],
+        columns: [
+          _column_192,
+          _column_193,
+          _column_194,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape32 signalPreKeyStores = Shape32(
+      source: i0.VersionedTable(
+        entityName: 'signal_pre_key_stores',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(pre_key_id)',
+        ],
+        columns: [
+          _column_195,
+          _column_196,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape11 signalSenderKeyStores = Shape11(
+      source: i0.VersionedTable(
+        entityName: 'signal_sender_key_stores',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(sender_key_name)',
+        ],
+        columns: [
+          _column_197,
+          _column_198,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape33 signalSessionStores = Shape33(
+      source: i0.VersionedTable(
+        entityName: 'signal_session_stores',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(device_id, name)',
+        ],
+        columns: [
+          _column_192,
+          _column_193,
+          _column_199,
+          _column_118,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape34 messageActions = Shape34(
+      source: i0.VersionedTable(
+        entityName: 'message_actions',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(message_id, contact_id, type)',
+        ],
+        columns: [
+          _column_174,
+          _column_183,
+          _column_144,
+          _column_200,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+  late final Shape35 groupHistories = Shape35(
+      source: i0.VersionedTable(
+        entityName: 'group_histories',
+        withoutRowId: false,
+        isStrict: false,
+        tableConstraints: [
+          'PRIMARY KEY(group_history_id)',
+        ],
+        columns: [
+          _column_201,
+          _column_158,
+          _column_202,
+          _column_203,
+          _column_204,
+          _column_205,
+          _column_206,
+          _column_144,
+          _column_200,
+        ],
+        attachedDatabase: database,
+      ),
+      alias: null);
+}
+
+class Shape36 extends i0.VersionedTable {
+  Shape36({required super.source, required super.alias}) : super.aliased();
+  i1.GeneratedColumn<String> get mediaId =>
+      columnsByName['media_id']! as i1.GeneratedColumn<String>;
+  i1.GeneratedColumn<String> get type =>
+      columnsByName['type']! as i1.GeneratedColumn<String>;
+  i1.GeneratedColumn<String> get uploadState =>
+      columnsByName['upload_state']! as i1.GeneratedColumn<String>;
+  i1.GeneratedColumn<String> get downloadState =>
+      columnsByName['download_state']! as i1.GeneratedColumn<String>;
+  i1.GeneratedColumn<int> get requiresAuthentication =>
+      columnsByName['requires_authentication']! as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<int> get stored =>
+      columnsByName['stored']! as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<int> get isDraftMedia =>
+      columnsByName['is_draft_media']! as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<int> get preProgressingProcess =>
+      columnsByName['pre_progressing_process']! as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<String> get reuploadRequestedBy =>
+      columnsByName['reupload_requested_by']! as i1.GeneratedColumn<String>;
+  i1.GeneratedColumn<int> get displayLimitInMilliseconds =>
+      columnsByName['display_limit_in_milliseconds']!
+          as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<int> get removeAudio =>
+      columnsByName['remove_audio']! as i1.GeneratedColumn<int>;
+  i1.GeneratedColumn<i2.Uint8List> get downloadToken =>
+      columnsByName['download_token']! as i1.GeneratedColumn<i2.Uint8List>;
+  i1.GeneratedColumn<i2.Uint8List> get encryptionKey =>
+      columnsByName['encryption_key']! as i1.GeneratedColumn<i2.Uint8List>;
+  i1.GeneratedColumn<i2.Uint8List> get encryptionMac =>
+      columnsByName['encryption_mac']! as i1.GeneratedColumn<i2.Uint8List>;
+  i1.GeneratedColumn<i2.Uint8List> get encryptionNonce =>
+      columnsByName['encryption_nonce']! as i1.GeneratedColumn<i2.Uint8List>;
+  i1.GeneratedColumn<i2.Uint8List> get storedFileHash =>
+      columnsByName['stored_file_hash']! as i1.GeneratedColumn<i2.Uint8List>;
+  i1.GeneratedColumn<int> get createdAt =>
+      columnsByName['created_at']! as i1.GeneratedColumn<int>;
+}
+
+i1.GeneratedColumn<int> _column_207(String aliasedName) =>
+    i1.GeneratedColumn<int>('pre_progressing_process', aliasedName, true,
+        type: i1.DriftSqlType.int, $customConstraints: 'NULL');
 i0.MigrationStepWithVersion migrationSteps({
   required Future<void> Function(i1.Migrator m, Schema2 schema) from1To2,
   required Future<void> Function(i1.Migrator m, Schema3 schema) from2To3,
@@ -4311,6 +4696,7 @@ i0.MigrationStepWithVersion migrationSteps({
   required Future<void> Function(i1.Migrator m, Schema6 schema) from5To6,
   required Future<void> Function(i1.Migrator m, Schema7 schema) from6To7,
   required Future<void> Function(i1.Migrator m, Schema8 schema) from7To8,
+  required Future<void> Function(i1.Migrator m, Schema9 schema) from8To9,
 }) {
   return (currentVersion, database) async {
     switch (currentVersion) {
@@ -4349,6 +4735,11 @@ i0.MigrationStepWithVersion migrationSteps({
         final migrator = i1.Migrator(database, schema);
         await from7To8(migrator, schema);
         return 8;
+      case 8:
+        final schema = Schema9(database: database);
+        final migrator = i1.Migrator(database, schema);
+        await from8To9(migrator, schema);
+        return 9;
       default:
         throw ArgumentError.value('Unknown migration from $currentVersion');
     }
@@ -4363,6 +4754,7 @@ i1.OnUpgrade stepByStep({
   required Future<void> Function(i1.Migrator m, Schema6 schema) from5To6,
   required Future<void> Function(i1.Migrator m, Schema7 schema) from6To7,
   required Future<void> Function(i1.Migrator m, Schema8 schema) from7To8,
+  required Future<void> Function(i1.Migrator m, Schema9 schema) from8To9,
 }) =>
     i0.VersionedSchema.stepByStepHelper(
         step: migrationSteps(
@@ -4373,4 +4765,5 @@ i1.OnUpgrade stepByStep({
       from5To6: from5To6,
       from6To7: from6To7,
       from7To8: from7To8,
+      from8To9: from8To9,
     ));